+++ UmecaApp/com/umeca/entities/reviewer/RelSchedule.cs
@@ -11,19 +11,19 @@ namespace UmecaApp
 		}
 
 		[PrimaryKey, AutoIncrement, Column("id")]
-		public long Id{ get; set; }
+		public int Id{ get; set; }
 
 		[Column("imputed_home_id")]
-		public long ImputedHomeId{ get; set; }
+		public int ImputedHomeId{ get; set; }
 
 		[Column("job_id")]
-		public long JobId{ get; set; }
+		public int JobId{ get; set; }
 
 		[Column("school_id")]
-		public long SchoolId{ get; set; }
+		public int SchoolId{ get; set; }
 
 		[Column("schedule_id")]
-		public long ScheduleId{ get; set; }
+		public int ScheduleId{ get; set; }
 	}
 }
